pyproject.toml
[project]
name = "object-state"
version = "3.0.6"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.7.4",
    "asyncpg==0.30.0",
    "cryptography==44.0.3",
    "fastapi==0.115.2",
    "grpcio==1.64.1",
    "numpy==1.26.4",
    "pandas==2.0.3",
    "protobuf==5.29.3",
    "pydantic==2.11.4",
    "pydantic-settings==2.9.1",
    "pyjwt==2.10.1",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    "sqlalchemy[asyncio]==2.0.41",
    "uvicorn==0.34.2",
]

[dependency-groups]
dev = [
]
lint = [
    "ruff==0.11.9",
]
tests = [
]
security = [
    "pip-audit==2.7.3",
]
migrations = [
    "alembic==1.7.4",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.41",
]
